Refactoring A simple implementation of EstimateDeliveryTime and EstimatePriceCalculation they have a lot of TODO

+++ model/src/main/java/com/alevel/deliverit/billing/Money.java
@@ -6,15 +6,15 @@ import java.util.Currency;
  * @author Vadym Mitin
  */
 public class Money {
-    private final int value;
+    private final long value;
     private final Currency currency;
 
-    public Money(int value, Currency currency) {
+    public Money(long value, Currency currency) {
         this.value = value;
         this.currency = currency;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
